+++ runtime/dex_file.h
@@ -471,7 +471,7 @@ class DexFile {
   static bool IsVersionValid(const uint8_t* magic);
 
   // Returns the number of string identifiers in the .dex file.
-  size_t NumStringIds() const {
+  uint32_t NumStringIds() const {
     DCHECK(header_ != nullptr) << GetLocation();
     return header_->string_ids_size_;
   }
@@ -485,7 +485,7 @@ class DexFile {
   uint32_t GetIndexForStringId(const StringId& string_id) const {
     CHECK_GE(&string_id, string_ids_) << GetLocation();
     CHECK_LT(&string_id, string_ids_ + header_->string_ids_size_) << GetLocation();
-    return &string_id - string_ids_;
+    return static_cast<uint32_t>(&string_id - string_ids_);
   }
 
   int32_t GetStringLength(const StringId& string_id) const;
@@ -561,7 +561,7 @@ class DexFile {
   const TypeId* FindTypeId(uint32_t string_idx) const;
 
   // Returns the number of field identifiers in the .dex file.
-  size_t NumFieldIds() const {
+  uint32_t NumFieldIds() const {
     DCHECK(header_ != nullptr) << GetLocation();
     return header_->field_ids_size_;
   }
@@ -575,7 +575,7 @@ class DexFile {
   uint32_t GetIndexForFieldId(const FieldId& field_id) const {
     CHECK_GE(&field_id, field_ids_) << GetLocation();
     CHECK_LT(&field_id, field_ids_ + header_->field_ids_size_) << GetLocation();
-    return &field_id - field_ids_;
+    return static_cast<uint32_t>(&field_id - field_ids_);
   }
 
   // Looks up a field by its declaring class, name and type
@@ -601,7 +601,7 @@ class DexFile {
   }
 
   // Returns the number of method identifiers in the .dex file.
-  size_t NumMethodIds() const {
+  uint32_t NumMethodIds() const {
     DCHECK(header_ != nullptr) << GetLocation();
     return header_->method_ids_size_;
   }
@@ -615,7 +615,7 @@ class DexFile {
   uint32_t GetIndexForMethodId(const MethodId& method_id) const {
     CHECK_GE(&method_id, method_ids_) << GetLocation();
     CHECK_LT(&method_id, method_ids_ + header_->method_ids_size_) << GetLocation();
-    return &method_id - method_ids_;
+    return static_cast<uint32_t>(&method_id - method_ids_);
   }
 
   // Looks up a method by its declaring class, name and proto_id
@@ -665,7 +665,7 @@ class DexFile {
   uint16_t GetIndexForClassDef(const ClassDef& class_def) const {
     CHECK_GE(&class_def, class_defs_) << GetLocation();
     CHECK_LT(&class_def, class_defs_ + header_->class_defs_size_) << GetLocation();
-    return &class_def - class_defs_;
+    return static_cast<uint16_t>(&class_def - class_defs_);
   }
 
   // Returns the class descriptor string of a class definition.
@@ -727,12 +727,12 @@ class DexFile {
   uint16_t GetIndexForProtoId(const ProtoId& proto_id) const {
     CHECK_GE(&proto_id, proto_ids_) << GetLocation();
     CHECK_LT(&proto_id, proto_ids_ + header_->proto_ids_size_) << GetLocation();
-    return &proto_id - proto_ids_;
+    return static_cast<uint16_t>(&proto_id - proto_ids_);
   }
 
   // Looks up a proto id for a given return type and signature type list
   const ProtoId* FindProtoId(uint16_t return_type_idx,
-                             const uint16_t* signature_type_idxs, uint32_t signature_length) const;
+                             const uint16_t* signature_type_idxs, size_t signature_length) const;
   const ProtoId* FindProtoId(uint16_t return_type_idx,
                              const std::vector<uint16_t>& signature_type_idxs) const {
     return FindProtoId(return_type_idx, &signature_type_idxs[0], signature_type_idxs.size());
@@ -852,7 +852,7 @@ class DexFile {
     DISALLOW_COPY_AND_ASSIGN(LineNumFromPcContext);
   };
 
-  void InvokeLocalCbIfLive(void* context, int reg, uint32_t end_address,
+  void InvokeLocalCbIfLive(void* context, uint16_t reg, uint32_t end_address,
                            LocalInfo* local_in_reg, DexDebugNewLocalCb local_cb) const {
     if (local_cb != nullptr && local_in_reg[reg].is_live_) {
       local_cb(context, reg, local_in_reg[reg].start_address_, end_address,
