+++ runtime/oat_file-inl.h
@@ -35,7 +35,7 @@ inline uint32_t OatFile::OatMethod::GetOatQuickMethodHeaderOffset() const {
   if (method_header == nullptr) {
     return 0u;
   }
-  return reinterpret_cast<const uint8_t*>(method_header) - begin_;
+  return static_cast<uint32_t>(reinterpret_cast<const uint8_t*>(method_header) - begin_);
 }
 
 inline uint32_t OatFile::OatMethod::GetQuickCodeSizeOffset() const {
@@ -43,7 +43,8 @@ inline uint32_t OatFile::OatMethod::GetQuickCodeSizeOffset() const {
   if (method_header == nullptr) {
     return 0u;
   }
-  return reinterpret_cast<const uint8_t*>(&method_header->code_size_) - begin_;
+  return static_cast<uint32_t>(
+      reinterpret_cast<const uint8_t*>(&method_header->code_size_) - begin_);
 }
 
 inline size_t OatFile::OatMethod::GetFrameSizeInBytes() const {
@@ -92,7 +93,8 @@ inline uint32_t OatFile::OatMethod::GetGcMapOffsetOffset() const {
   if (method_header == nullptr) {
     return 0u;
   }
-  return reinterpret_cast<const uint8_t*>(&method_header->gc_map_offset_) - begin_;
+  return static_cast<uint32_t>(
+      reinterpret_cast<const uint8_t*>(&method_header->gc_map_offset_) - begin_);
 }
 
 inline uint32_t OatFile::OatMethod::GetMappingTableOffset() const {
@@ -105,7 +107,8 @@ inline uint32_t OatFile::OatMethod::GetMappingTableOffsetOffset() const {
   if (method_header == nullptr) {
     return 0u;
   }
-  return reinterpret_cast<const uint8_t*>(&method_header->mapping_table_offset_) - begin_;
+  return static_cast<uint32_t>(
+      reinterpret_cast<const uint8_t*>(&method_header->mapping_table_offset_) - begin_);
 }
 
 inline uint32_t OatFile::OatMethod::GetVmapTableOffset() const {
@@ -118,7 +121,8 @@ inline uint32_t OatFile::OatMethod::GetVmapTableOffsetOffset() const {
   if (method_header == nullptr) {
     return 0u;
   }
-  return reinterpret_cast<const uint8_t*>(&method_header->vmap_table_offset_) - begin_;
+  return static_cast<uint32_t>(
+      reinterpret_cast<const uint8_t*>(&method_header->vmap_table_offset_) - begin_);
 }
 
 inline const uint8_t* OatFile::OatMethod::GetMappingTable() const {
